+++ createtx.go
@@ -52,6 +52,8 @@ func (e InsufficientFunds) Error() string {
 		e.fee, e.in)
 }
 
+var UnsupportedTransactionType = errors.New("Only P2PKH transactions are supported")
+
 // ErrNonPositiveAmount represents an error where a bitcoin amount is
 // not positive (either negative, or zero).
 var ErrNonPositiveAmount = errors.New("amount is not positive")
@@ -139,20 +141,8 @@ func (w *Wallet) txToPairs(pairs map[string]btcutil.Amount,
 		amt += v
 	}
 
-	// Add outputs to new tx.
-	for addrStr, amt := range pairs {
-		addr, err := btcutil.DecodeAddress(addrStr, activeNet.Params)
-		if err != nil {
-			return nil, fmt.Errorf("cannot decode address: %s", err)
-		}
-
-		// Add output to spend amt to addr.
-		pkScript, err := btcscript.PayToAddrScript(addr)
-		if err != nil {
-			return nil, fmt.Errorf("cannot create txout script: %s", err)
-		}
-		txout := btcwire.NewTxOut(int64(amt), pkScript)
-		msgtx.AddTxOut(txout)
+	if err = addOutputs(msgtx, pairs); err != nil {
+		return nil, err
 	}
 
 	// Get current block's height and hash.
@@ -161,45 +151,10 @@ func (w *Wallet) txToPairs(pairs map[string]btcutil.Amount,
 		return nil, err
 	}
 
-	// Make a copy of msgtx before any inputs are added.  This will be
-	// used as a starting point when trying a fee and starting over with
-	// a higher fee if not enough was originally chosen.
-	txNoInputs := msgtx.Copy()
-
-	unspent, err := w.TxStore.UnspentOutputs()
+	eligible, err := w.findEligibleOuptuts(minconf, bs)
 	if err != nil {
 		return nil, err
 	}
-
-	// Filter out unspendable outputs, that is, remove those that (at this
-	// time) are not P2PKH outputs.  Other inputs must be manually included
-	// in transactions and sent (for example, using createrawtransaction,
-	// signrawtransaction, and sendrawtransaction).
-	eligible := make([]txstore.Credit, 0, len(unspent))
-	for i := range unspent {
-		switch btcscript.GetScriptClass(unspent[i].TxOut().PkScript) {
-		case btcscript.PubKeyHashTy:
-			if !unspent[i].Confirmed(minconf, bs.Height) {
-				continue
-			}
-			// Coinbase transactions must have have reached maturity
-			// before their outputs may be spent.
-			if unspent[i].IsCoinbase() {
-				target := btcchain.CoinbaseMaturity
-				if !unspent[i].Confirmed(target, bs.Height) {
-					continue
-				}
-			}
-
-			// Locked unspent outputs are skipped.
-			if w.LockedOutpoint(*unspent[i].OutPoint()) {
-				continue
-			}
-
-			eligible = append(eligible, unspent[i])
-		}
-	}
-
 	// Sort eligible inputs, as selectInputs expects these to be sorted
 	// by amount in reverse order.
 	sort.Sort(sort.Reverse(ByAmount(eligible)))
@@ -210,6 +165,11 @@ func (w *Wallet) txToPairs(pairs map[string]btcutil.Amount,
 	var changeAddr btcutil.Address
 	var changeIdx int
 
+	// Make a copy of msgtx before any inputs are added.  This will be
+	// used as a starting point when trying a fee and starting over with
+	// a higher fee if not enough was originally chosen.
+	txNoInputs := msgtx.Copy()
+
 	// Get the number of satoshis to increment fee by when searching for
 	// the minimum tx fee needed.
 	fee := btcutil.Amount(0)
@@ -218,7 +178,7 @@ func (w *Wallet) txToPairs(pairs map[string]btcutil.Amount,
 		changeIdx = -1
 
 		// Select eligible outputs to be used in transaction based on the amount
-		// neededing to sent, and the current fee estimation.
+		// needed to be sent, and the current fee estimation.
 		inputs, btcin, err := selectInputs(eligible, amt, fee, minconf)
 		if err != nil {
 			return nil, err
@@ -257,41 +217,8 @@ func (w *Wallet) txToPairs(pairs map[string]btcutil.Amount,
 			changeIdx = int(r)
 		}
 
-		// Selected unspent outputs become new transaction's inputs.
-		for _, ip := range inputs {
-			msgtx.AddTxIn(btcwire.NewTxIn(ip.OutPoint(), nil))
-		}
-		for i, input := range inputs {
-			// Errors don't matter here, as we only consider the
-			// case where len(addrs) == 1.
-			_, addrs, _, _ := input.Addresses(activeNet.Params)
-			if len(addrs) != 1 {
-				continue
-			}
-			apkh, ok := addrs[0].(*btcutil.AddressPubKeyHash)
-			if !ok {
-				continue // don't handle inputs to this yet
-			}
-
-			ai, err := w.KeyStore.Address(apkh)
-			if err != nil {
-				return nil, fmt.Errorf("cannot get address info: %v", err)
-			}
-
-			pka := ai.(keystore.PubKeyAddress)
-
-			privkey, err := pka.PrivKey()
-			if err != nil {
-				return nil, fmt.Errorf("cannot get private key: %v", err)
-			}
-
-			sigscript, err := btcscript.SignatureScript(msgtx, i,
-				input.TxOut().PkScript, btcscript.SigHashAll, privkey,
-				ai.Compressed())
-			if err != nil {
-				return nil, fmt.Errorf("cannot create sigscript: %s", err)
-			}
-			msgtx.TxIn[i].SignatureScript = sigscript
+		if err = w.addInputsToTx(msgtx, inputs); err != nil {
+			return nil, err
 		}
 
 		noFeeAllowed := false
@@ -306,21 +233,8 @@ func (w *Wallet) txToPairs(pairs map[string]btcutil.Amount,
 		}
 	}
 
-	// Validate msgtx before returning the raw transaction.
-	flags := btcscript.ScriptCanonicalSignatures
-	bip16 := time.Now().After(btcscript.Bip16Activation)
-	if bip16 {
-		flags |= btcscript.ScriptBip16
-	}
-	for i, txin := range msgtx.TxIn {
-		engine, err := btcscript.NewScript(txin.SignatureScript,
-			selectedInputs[i].TxOut().PkScript, i, msgtx, flags)
-		if err != nil {
-			return nil, fmt.Errorf("cannot create script engine: %s", err)
-		}
-		if err = engine.Execute(); err != nil {
-			return nil, fmt.Errorf("cannot validate transaction: %s", err)
-		}
+	if err = validateMsgTx(msgtx, selectedInputs); err != nil {
+		return nil, err
 	}
 
 	buf := bytes.Buffer{}
@@ -336,6 +250,119 @@ func (w *Wallet) txToPairs(pairs map[string]btcutil.Amount,
 		changeIndex: changeIdx,
 	}
 	return info, nil
+}
+
+func addOutputs(msgtx *btcwire.MsgTx, pairs map[string]btcutil.Amount) error {
+	for addrStr, amt := range pairs {
+		addr, err := btcutil.DecodeAddress(addrStr, activeNet.Params)
+		if err != nil {
+			return fmt.Errorf("cannot decode address: %s", err)
+		}
+
+		// Add output to spend amt to addr.
+		pkScript, err := btcscript.PayToAddrScript(addr)
+		if err != nil {
+			return fmt.Errorf("cannot create txout script: %s", err)
+		}
+		txout := btcwire.NewTxOut(int64(amt), pkScript)
+		msgtx.AddTxOut(txout)
+	}
+	return nil
+}
+
+func (w *Wallet) findEligibleOuptuts(minconf int, bs *keystore.BlockStamp) ([]txstore.Credit, error) {
+	unspent, err := w.TxStore.UnspentOutputs()
+	if err != nil {
+		return nil, err
+	}
+	// Filter out unspendable outputs, that is, remove those that (at this
+	// time) are not P2PKH outputs.  Other inputs must be manually included
+	// in transactions and sent (for example, using createrawtransaction,
+	// signrawtransaction, and sendrawtransaction).
+	eligible := make([]txstore.Credit, 0, len(unspent))
+	for i := range unspent {
+		switch btcscript.GetScriptClass(unspent[i].TxOut().PkScript) {
+		case btcscript.PubKeyHashTy:
+			if !unspent[i].Confirmed(minconf, bs.Height) {
+				continue
+			}
+			// Coinbase transactions must have have reached maturity
+			// before their outputs may be spent.
+			if unspent[i].IsCoinbase() {
+				target := btcchain.CoinbaseMaturity
+				if !unspent[i].Confirmed(target, bs.Height) {
+					continue
+				}
+			}
+
+			// Locked unspent outputs are skipped.
+			if w.LockedOutpoint(*unspent[i].OutPoint()) {
+				continue
+			}
+
+			eligible = append(eligible, unspent[i])
+		}
+	}
+	return eligible, nil
+}
+
+// For every unspent output given, add a new input to the given MsgTx. Only P2PKH outputs are
+// supported at this point.
+func (w *Wallet) addInputsToTx(msgtx *btcwire.MsgTx, outputs []txstore.Credit) error {
+	for _, ip := range outputs {
+		msgtx.AddTxIn(btcwire.NewTxIn(ip.OutPoint(), nil))
+	}
+	for i, output := range outputs {
+		// Errors don't matter here, as we only consider the
+		// case where len(addrs) == 1.
+		_, addrs, _, _ := output.Addresses(activeNet.Params)
+		if len(addrs) != 1 {
+			continue
+		}
+		apkh, ok := addrs[0].(*btcutil.AddressPubKeyHash)
+		if !ok {
+			return UnsupportedTransactionType
+		}
+
+		ai, err := w.KeyStore.Address(apkh)
+		if err != nil {
+			return fmt.Errorf("cannot get address info: %v", err)
+		}
+
+		pka := ai.(keystore.PubKeyAddress)
+
+		privkey, err := pka.PrivKey()
+		if err != nil {
+			return fmt.Errorf("cannot get private key: %v", err)
+		}
+
+		sigscript, err := btcscript.SignatureScript(
+			msgtx, i, output.TxOut().PkScript, btcscript.SigHashAll, privkey, ai.Compressed())
+		if err != nil {
+			return fmt.Errorf("cannot create sigscript: %s", err)
+		}
+		msgtx.TxIn[i].SignatureScript = sigscript
+	}
+	return nil
+}
+
+func validateMsgTx(msgtx *btcwire.MsgTx, inputs []txstore.Credit) error {
+	flags := btcscript.ScriptCanonicalSignatures
+	bip16 := time.Now().After(btcscript.Bip16Activation)
+	if bip16 {
+		flags |= btcscript.ScriptBip16
+	}
+	for i, txin := range msgtx.TxIn {
+		engine, err := btcscript.NewScript(
+			txin.SignatureScript, inputs[i].TxOut().PkScript, i, msgtx, flags)
+		if err != nil {
+			return fmt.Errorf("cannot create script engine: %s", err)
+		}
+		if err = engine.Execute(); err != nil {
+			return fmt.Errorf("cannot validate transaction: %s", err)
+		}
+	}
+	return nil
 }
 
 // minimumFee calculates the minimum fee required for a transaction.
